Change my-tools path upon select
